requirements.txt
pyyaml>=6.0
torch>=2.0
torchvision>=0.15
numpy
pandas
matplotlib
scikit-learn
Pillow
tqdm
scipy
h5py
